code refactoring about compute abstraction
