+++ pywren_ibm_cloud/compute/backends/ibm_cf/__init__.py
@@ -1,0 +1,1 @@
+from .ibm_cf import ComputeBackend
