+++ pywren_ibm_cloud/storage/__init__.py
@@ -1,1 +1,1 @@
-from .internal_storage import InternalStorage
+from .storage import InternalStorage
