+++ pywren_ibm_cloud/storage/backends/ibm_cos/__init__.py
@@ -1,0 +1,1 @@
+from .ibm_cos import StorageBackend
