+++ pywren_ibm_cloud/storage/backends/swift/__init__.py
@@ -1,0 +1,1 @@
+from .swift import StorageBackend
